Create CallMonitorHttpService
The service is created as a foreground service so that it can run even after device restarts without necessarily having to open the call monitor app to start it. Also following Android's background execution requirements, it must display a notification so that the user is aware of a running service.

Since the service is started while the app is in the background, a notification is required since Android 8.0.