Display Call Logs in Client App
Create ViewModel to manage fetching of the call logs and Ui state
Create a CallMonitor public api. The fetchCallLogs() public method needs to be called only after successfully binding to the service.
The CallLogsUiContract defines the contract with which the ui communicates with the view model and also defines the uiState.