Fix service crash after reboot
The http service would unreliably crash on device restart due to incomplete koin initialization.
This attempts to fix the issue by removing the unnecessary 'isRunning' in the onCreate method, to ensure that the HttpService always gets properly initialized when the service is first created.
This also introduces a courotineScope with which to execute the http service in.
The 'BootCompletedReceiver' class is moved to the 'receiver' package for better visibility.